Move sleep to the end of frame on iOS to fix stuttering
The main loop was

while (running) {
    sleep
    step
    draw
    swap
}

so the sleep happened immediately after the buffer swap. But on iOS
there's no main loop and it was

- (void) drawView {
    sleep
    step
    draw
    swap
}

which caused LOTS of dropped frame because we were sleeping at the
beginning of each frame with information of the previous one. Whenever
a frame took longer than the last one this would likely cause a frame
drop.